feat(dpav-1441): added generic component for add entity screens
