+++ frontend/src/components/AppWrapper.tsx
@@ -23,11 +23,11 @@ import IncidentPickerPage from '../pages/IncidentPickerPage';
 import Settings from '../pages/Settings';
 import Tasks from '../pages/Tasks';
 import Layout from './Layout';
-import LogbookV2 from '../pages/LogbookV2';
 import MyProfile from '../pages/MyProfile';
 import AdminUserList from '../pages/AdminUserList';
 import AdminViewUser from '../pages/AdminViewUser';
 import AdminNewUser from '../pages/AdminNewUser';
+import { CreateLogEntry } from '../pages/CreateLogEntry';
 
 const AppWrapper = () => {
   const router = createBrowserRouter([
@@ -42,7 +42,7 @@ const AppWrapper = () => {
         { path: 'createlog', element: <CreateLog /> },
         { path: 'incident/:incidentId', element: <Overview /> },
         { path: 'logbook/:incidentId', element: <Logbook /> },
-        { path: 'logbook/v2/:incidentId', element: <LogbookV2 /> },
+        { path: 'logbook/:incidentId/create', element: <CreateLogEntry /> },
         { path: 'tasks', element: <Tasks /> },
         { path: 'tasks/:incidentId', element: <IncidentTasks /> },
         { path: 'incidents/pick', element: <IncidentPickerPage /> },
